Programmed the web app to incorporate time.
Time is included in the index.html and main.js. As for the main.js, time will be used to display a moon if the weather is clear after 6 pm and before 6 am. For the HTML file, time will be used to choose a CSS file: one that will have a daytime background and one that will have a nighttime background.